Move deploy to component
