+++ src/views/Home.tsx
@@ -1,14 +1,12 @@
-import React, { Component } from "react";
-import { Button } from "react-bootstrap";
-import DeployVCSim from "../components/deployment/VCSim";
-import DeployForm from "../components/deployment/Form";
+import React from 'react';
+import DeployForm from '../components/deployment/Form';
 
 export default function Home() {
-    return (
-      <main role="main" className="container">
-        <h2>Home</h2>
-        <br />
-        {DeployForm()}
-      </main>
-    );
+  return (
+    <main role="main" className="container">
+      <h2>Home</h2>
+      <br />
+      <DeployForm />
+    </main>
+  );
 }
